removed urls and renamed model file
