+++ django_reference_app/django_reference_app/modelforms.py
@@ -1,5 +1,5 @@
 from django import forms
-from django_reference_app.dbmodels import Person
+from django_reference_app.models import Person
 
 class tableForms(forms.ModelForm):
     class Meta:
