+++ django_reference_app/django_reference_app/views.py
@@ -12,5 +12,5 @@ def showform(request):
 
     context = {'name': name, 'job': job}
         
-    return render(request, 'clusterwebtwo.html', context)
+    return render(request, 'infopage.html', context)
